Move case utils tests
