+++ custom/enikshay/tests/test_case_utils.py
@@ -1,0 +1,61 @@
+from datetime import datetime
+import pytz
+
+from django.test import TestCase
+from custom.enikshay.tests.utils import ENikshayCaseStructureMixin
+from corehq.form_processor.tests.utils import run_with_all_backends
+
+from custom.enikshay.case_utils import (
+    get_open_episode_case_from_person,
+    get_adherence_cases_between_dates,
+)
+
+
+class ENikshayCaseUtilsTests(ENikshayCaseStructureMixin, TestCase):
+    @run_with_all_backends
+    def test_get_adherence_cases_between_dates(self):
+        self.create_case_structure()
+        adherence_dates = [
+            datetime(2005, 7, 10),
+            datetime(2016, 8, 10),
+            datetime(2016, 8, 11),
+            datetime(2016, 8, 12),
+            datetime(2017, 9, 10),
+        ]
+        self.create_adherence_cases(adherence_dates)
+        fetched_cases = get_adherence_cases_between_dates(
+            self.domain,
+            self.person_id,
+            start_date=datetime(2016, 7, 1, tzinfo=pytz.UTC),
+            end_date=datetime(2016, 8, 13, tzinfo=pytz.UTC),
+        )
+        self.assertEqual(len(fetched_cases), 3)
+        self.assertItemsEqual(
+            ["2016-08-10", "2016-08-11", "2016-08-12"],
+            [case.case_id for case in fetched_cases]
+        )
+
+        fetched_cases = get_adherence_cases_between_dates(
+            self.domain,
+            self.person_id,
+            start_date=datetime(2010, 7, 1, tzinfo=pytz.UTC),
+            end_date=datetime(2010, 8, 13, tzinfo=pytz.UTC),
+        )
+        self.assertEqual(len(fetched_cases), 0)
+
+        fetched_cases = get_adherence_cases_between_dates(
+            self.domain,
+            self.person_id,
+            start_date=datetime(2016, 8, 10, tzinfo=pytz.UTC),
+            end_date=datetime(2016, 8, 10, tzinfo=pytz.UTC),
+        )
+        self.assertEqual(len(fetched_cases), 1)
+        self.assertEqual(
+            "2016-08-10",
+            fetched_cases[0].case_id,
+        )
+
+    @run_with_all_backends
+    def test_get_episode(self):
+        self.create_case_structure()
+        self.assertEqual(get_open_episode_case_from_person(self.domain, 'person').case_id, 'episode')
